[mc-plugin] Toggle account Enabled state when modems disappear. Fixes JB#34135
With the old implementation there was race issue when modems were
appearing and disappearing quickly. When for example /ril_0 modem was
removed, the operation to actually remove the account and accompanying
DBus object path took so long that it was possible that /ril_0 appeared
in Ofono again. Then when mc-plugin started to create new account for
the modem mission-control would crash as the old object path wasn't
removed soon enough. As it is quite difficult to monitor when the
account object is really completely torn down, easier solution is to
toggle account's Enabled parameter instead of removing and adding the
account entirely.

So, to fix the race issue crash, when modem first appears we create a
new account for the modem with Enabled true, and whenever modem with
accompanying account (dis)appears, we toggle Enabled parameter.